+++ src/pages/Main/Main.js
@@ -83,7 +83,7 @@ export default function Main() {
 
     return (
         <>
-            <LoadingScreen {...{ loading, dataReadError }} />
+            <LoadingScreen {...{ loading, dataReadError, userState, userDispatch }} />
             <BurgerMenu {...{ menuOpen, setMenuOpen, currentPage, setCurrentPage }} />
             <Title />
             {!loading &&
@@ -196,9 +196,7 @@ function userReducer(state, action) {
             }
             let result = action.payload.settings.consume
                 ? data_utils(action.payload.data).project.consume(action.payload.project, state.inventory,
-                    state.settings.use_legacy ? _CONSTANTS.RECIPE_NOTE.LEGACY
-                        : state.settings.use_legacy_2 ? _CONSTANTS.RECIPE_NOTE.LEGACY_2
-                        : undefined,
+                    state.settings.region,
                     action.payload.project.details.ignored_rarity || {})
                 : state.inventory;
             if (action.payload.settings.save && action.payload.project.type === "character") {