+++ src/pages/Main/Settings.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import Alert from '@mui/material/Alert';
 import Button from '@mui/material/Button';
 import Card from '@mui/material/Card';
@@ -7,89 +7,69 @@ import CardContent from '@mui/material/CardContent';
 import Typography from '@mui/material/Typography';
 import Grid from '@mui/material/Grid';
 import Switch from '@mui/material/Switch';
+import Select from '@mui/material/Select';
+import InputLabel from '@mui/material/InputLabel';
+import MenuItem from '@mui/material/MenuItem';
+import FormControl from '@mui/material/FormControl';
 
 export default function Settings({ userState, userDispatch, hidden }) {
-    const [legacy, setLegacy] = useState(userState.settings.use_legacy);
-    const [legacy2, setLegacy2] = useState(userState.settings.use_legacy_2);
+    const [region, setRegion] = useState(userState.settings.region || 'JP');
+
+    useEffect(() => {
+        if (!userState.settings.region || userState.settings.region === "") {
+            return;
+        }
+        if (region !== userState.settings.region) {
+            setRegion(userState.settings.region);
+        }
+    }, [userState.settings.region]);
+
+    function handleChange(event) {
+        setRegion(event.target.value);
+        userDispatch({
+            type: 'SET_SETTINGS',
+            payload: {
+                key: "region",
+                data: event.target.value
+            },
+        });
+    }
     return (
         <Grid container direction="column" alignItems="center" justifyContent="center" gap={2} hidden={hidden}>
             <Alert severity="info" className="w-[75vw]">
                 Quest related settings (quest range, sort options, item filter) are located in the <strong>
-                Quest Drawer</strong>.<br /> Enable some projects to be able to open the quest drawer and
+                Quest Drawer</strong>.<br />Enable some projects to be able to open the quest drawer and
                 edit quest related settings.
             </Alert>
             <Card sx={{ width: "75vw", padding: 2 }}>
                 <CardHeader
-                    title={<Typography variant="h6">Use Legacy Version</Typography>}
-                    subheader={"Reduce cost of some silver, gold, and purple rarity items."}
-                    action={<Switch checked={legacy}
-                        aria-label="Use Legacy Version"
-                        onClick={() => {
-                            if (!userState.settings.use_legacy && legacy2) {
-                                setLegacy2(false);
-                                userDispatch({
-                                    type: "SET_SETTINGS",
-                                    payload: {
-                                        key: "use_legacy_2",
-                                        data: false
-                                    },
-                                });
-                            }
-                            setLegacy(!userState.settings.use_legacy);
-                            userDispatch({
-                                type: "SET_SETTINGS",
-                                payload: {
-                                    key: "use_legacy",
-                                    data: !userState.settings.use_legacy
-                                },
-                            });
-                    }} />}
+                    title={<Typography variant="h6">Game Region</Typography>}
+                    subheader={"Change character names, equipment names, and equipment recipe used."}
+                    action={
+                    <FormControl className="mt-2">
+                        <InputLabel id="region-setting-label">Game Region</InputLabel>
+                        <Select
+                            labelId="region-setting-label"
+                            label="Game Region"
+                            value={region}
+                            onChange={handleChange}
+                        >
+                            <MenuItem value={"JP"}>Japan (JP)</MenuItem>
+                            <MenuItem value={"CN"}>China (CN)</MenuItem>
+                            <MenuItem value={"EN"}>English (EN)</MenuItem>
+                            <MenuItem value={"KR"}>Korea (KR)</MenuItem>
+                            <MenuItem value={"TW"}>Taiwan (TW)</MenuItem>
+                        </Select>
+                    </FormControl>}
                 />
                 <CardContent>
                     <Typography variant="body2" color="text.secondary">
-                        As of <strong>August 30, 2019</strong>, the Japanese server released an update to reduce the
-                        amount of items necessary to build some items. <strong className="italic">Servers using an
-                        older version of the game may not have this change yet</strong>, so this setting must be
-                        enabled to use the legacy recipe of some items. This setting can not be used together with
-                        "Use Legacy Version 2".
-                    </Typography>
-                </CardContent>
-            </Card>
-            <Card sx={{ width: "75vw", padding: 2 }}>
-                <CardHeader
-                    title={<Typography variant="h6">Use Legacy Version 2</Typography>}
-                    subheader={"Reduce cost of some silver, gold, purple, red, and green rarity items."}
-                    action={<Switch checked={legacy2}
-                        aria-label="Use Legacy Version 2"
-                        onClick={() => {
-                            if (!userState.settings.use_legacy_2 && legacy) {
-                                // use_legacy_2 from false to true
-                                console.log("legacy 2 from false to true");
-                                setLegacy(false);
-                                userDispatch({
-                                    type: "SET_SETTINGS",
-                                    payload: {
-                                        key: "use_legacy",
-                                        data: false
-                                    },
-                                });
-                            }
-                            setLegacy2(!userState.settings.use_legacy_2);
-                            userDispatch({
-                                type: "SET_SETTINGS",
-                                payload: {
-                                    key: "use_legacy_2",
-                                    data: !userState.settings.use_legacy_2
-                                },
-                            });
-                    }} />}
-                />
-                <CardContent>
-                    <Typography variant="body2" color="text.secondary">
-                        As of <strong>February 28, 2022</strong>, the Japanese server released yet another equipment
-                        cost reduction update. <strong className="italic">Servers using an older version of the game
-                        may not have this change yet</strong>, so this setting must be enabled to use the legacy_2
-                        recipe of some items. This setting can not be used together with "Use Legacy Version".
+                        <code>priconne-quest-helper</code> is based off the Japan version of the game, but each game
+                        region is currently at a different state. This means that equipment costs may be different or
+                        may use different costs at certain periods of the game. <strong>Make sure this setting
+                        matches your game region so that equipment calculations are as accurate as possible.</strong>
+                        <br /><br />
+                        If a certain equipment does not exist yet, the current Japanese recipe will be used as a fallback.
                     </Typography>
                 </CardContent>
             </Card>